Issue#1061 finalize individual history decimal point (#1077)
* Refactor the LeagueAverage from int to double

* Update client display version

+++ NineTapTour/NineTapTour/Models/FinalizeTemp.cs
@@ -32,7 +32,7 @@ namespace NineTapTour.Models
         public bool UseGame3 { get; set; }
         [DefaultValue(true)]
         public bool UseGame4 { get; set; }
-        public int LeagueAverage { get; set; }
+        public double LeagueAverage { get; set; }
         public int AdjustedAvg { get; set; }
         public string Notes { get; set; }
         public int ScratchTotal { get; set; }
